2-23-25 - Major mechanic adjustment
Changed order of operations so all seals become active at the beginning of a turn. This is done to reduce emphasis on the speed stat with regards to seals - anyone who moves last in a turn, be it a boss or fighter, will see less value in sealing, as the opposing side can now always react to it.