8-28-25 -  Visual Adjustments
Retooled origin points for player sprites to be more logical for a top down view.
Adusted visual aspects of escalator. This is all still temp of course, but it feels more natural now and is consistent with the rest of the world.
Fixed issue where floor change trigger would crash the game if no teleport target is assigned.