10-5-25 - More Flee Mechanics
Fleeing a battle is now subject to an RNG chance based on the fastest enemy's speed relative to the fastest player's speed.
If a battle is successfully fled, the enemies from that encounter will be stashed and have a chance of reappearing in subsequent battles. If the player wins a battle, stashed enemies will reset.