10-21-25 - Slime Adjustments
Slimes will now prioritize attacking if duplicate is sealed. This will make it easier to acquire their attack.
Reduced base power of Slime attack to 35
Duplicate no longer fails if number of active entities is 4.
Duplicate now outputs a failure message if the effect fails to activate.
Slimes now spawn in groups of 2 to 3.
Fixed issue where HP and MP bars would appear to be empty when they weren't due to the border