12-3-25 - Minor Adjustments and Fixes
Reduced Slime duplication chance from 0.8 to 0.65
Level Up message is now in color
Added turn seeds to sort out ties, patching out a really absurd engine limitation. Not sure if this actually caused problems or not.
L group enemies can no longer be higher leveled than 2 if anyone in the party is level 1. I don't see this staying in permanently, so it's toggled with a constant.
Fixed issue where certain kinds of skills would bypass normal disable behaviors, and would appear to be allowed to be sealed more than once.

It feels so good to be back doing stuff like this again...