12-8-25 - Enemy Seal Fix
For reasons unknown to all that may or may not be watching, the seal index was being randomized out of all the options. Evan, why would you think that's a good idea? It just prevents coherent boss strategies from existing? Perhaps I could see a use case for it, but that'd be for specific kinds of enemies, NOT all of them. Please just use -1 as a wildcard if we are actually to do this though.

Anyway this has been fixed.